move plotting functions to new module
