+++ fetch3/results/results.py
@@ -5,8 +5,6 @@ Convenience functions:
 """
 import xarray as xr
 import pandas as pd
-import hvplot.pandas
-import hvplot.xarray
 import yaml
 
 from pathlib import Path
@@ -59,25 +57,6 @@ def load_met_data(filein, timevar='TIMESTAMP_START'):
     return met
 
 
-def plot_2d_slice(ds=None, z=None, var=None):
-
-    plot = ds[var].sel(z=z).hvplot.line(x='time')
-    return plot
-
-def plot_precip_and_swc(model_ds, z=None, obs=None, obs_P='P_F', obs_swc='SWC_F_MDS_1', scale_obs_swc=True):
-
-    obs_swc = obs[obs_swc]
-
-    # If obs are in % instead of fraction
-    if scale_obs_swc:
-        obs_swc = obs_swc / 100
-
-    precip_plot = obs[obs_P].hvplot.line(xlim=(model_ds.time.min().values, model_ds.time.max().values))
-    swc_obs_plot = obs_swc.hvplot.line(xlim=(model_ds.time.min().values, model_ds.time.max().values))
-    swc_model_plot = model_ds.THETA.sel(z=z).hvplot.line(x='time', label='model SWC')
-    swc_plot = (swc_obs_plot * swc_model_plot).opts(ylim=(0,0.3))
-    return (precip_plot + swc_plot).cols(1)
-
 def calc_canopy1d(res):
 
     dz = res.cfg.model_options.dz
@@ -107,10 +86,12 @@ def get_best_opt_results(exp_dir):
     best_trial_index = list(best_trial.keys())[0]
     dir_best_trial = exp_dir / str(best_trial_index).zfill(6)
     return best_trial, dir_best_trial
+
+
 class Results:
 
     def __init__(self, output_dir, opt=True, label=None, config_name=None, data_dir=None, obs_file=None, obs_tvar='TIMESTAMP'):
-        
+
         if opt:
             self.best_trial, self.dir_best_trial = get_best_opt_results(output_dir)
             self.output_dir = self.dir_best_trial
